refactor(desktop): isolate MCP inventory queries
